UPD refactor and expand sync/storage handlers
- Sync storage handlers were moved to main sync/storage package.
- Methods needed for batch job were added.
- Related: #7
